compare function works as intended, replaced everywhere. Started used containing/containingDeep on integration tests
