+++ src/argsComparisons/compareArgs3.spec.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
-import { hasher } from "../hasher";
 
-import { Parser, containing, partial, schema, containingDeep, partialDeep } from "../behaviours/constructs";
+import { containing, partial, schema, containingDeep, partialDeep } from "../behaviours/constructs";
 import { randomUUID } from "crypto";
+import { compare } from "./compare";
 
 
 it("should compare numbers", () => {
@@ -619,221 +619,3 @@ it("should accept schemas in containingDeep", () => {
     ).toBe(false);
 });
 
-// TODO: schemas in maps & sets & arrays
-
-function compare(actual: any, expected: any) {
-    if (typeof expected === "object" && expected !== null) {
-        // fyi, (typeof null) equals "object". I know. #javascript
-        const isSchema = Object.keys(expected).some(key => key.endsWith("mockit__isSchema"));
-        if (isSchema) {
-            return (expected.schema as Parser).safeParse(actual).success;
-        }
-
-        const isPartial = Object.keys(expected).some(key => key.endsWith("mockit__isPartial"));
-        if (isPartial) {
-            if (Array.isArray(expected.original)) {
-                return actual?.every((item) => {
-                    return expected.original?.some(expectedItem => compare(item, expectedItem));
-                });
-            }
-
-            if (expected.original instanceof Map) {
-                return Array.from((actual as Map<any, any> ?? [])?.entries()).every(([key, value]) => {
-                    return compare(value, expected.original.get(key));
-                });
-            }
-
-            if (expected.original instanceof Set) {
-                return Array.from((actual as Set<any> ?? [])?.values()).every(value => {
-                    return Array.from(expected.original?.values()).some(expectedValue => compare(value, expectedValue));
-                });
-            }
-
-            return Object.keys(actual).every(key => {
-                console.log("partial: comparing values: ", actual[key], expected.original[key])
-                return compare(actual[key], expected.original[key]);
-            });
-        }
-
-        const isContaining = Object.keys(expected).some(key => key.endsWith("mockit__isContaining"));
-        if (isContaining) {
-            if (Array.isArray(expected.original)) {
-                return expected.original.every((item, index) => {
-                    return compare(actual[index], item);
-                });
-            }
-
-            if (expected.original instanceof Map) {
-                return Array.from(((expected?.original as Map<any, any>) ?? []).entries()).every(([key, value]) => {
-                    return compare(actual.get(key), value);
-                });
-            }
-
-            if (expected.original instanceof Set) {
-                return Array.from(((expected?.original as Set<any>) ?? []).values()).every(value => {
-                    return Array.from(actual.values()).some(actualValue => compare(actualValue, value));
-                });
-            }
-
-            return Object.keys(expected.original).every(key => {
-                return compare(actual[key], expected.original[key]);
-            });
-        }
-
-        const ispartialDeep = Object.keys(expected).some(key => key.endsWith("mockit__ispartialDeep"));
-        if (ispartialDeep) {
-            // Numbers, strings, etc.
-            if (typeof actual !== "object") {
-                return hasher.hash(actual) === hasher.hash(expected.original);
-            }
-
-            if (Array.isArray(expected.original)) {
-                return actual.every((item) => {
-                    return expected.original?.some(expectedItem => compare(item, partialDeep(expectedItem)));
-                });
-            }
-
-            if (expected.original instanceof Map) {
-                return Array.from(((actual as Map<any, any>) ?? [])?.entries()).every(([key, actualValue]) => {
-                    return compare(actualValue, partialDeep(expected.original.get(key)));
-                });
-            }
-
-            if (expected.original instanceof Set) {
-                return Array.from((actual as Set<any> ?? new Set()).values()).every(actualValue => {
-                    return Array.from(expected.original.values()).some(expectedValue => compare(actualValue, partialDeep(expectedValue)));
-                });
-            }
-            
-            // C'est reparti pour un tour
-            return Object.keys(actual).every(key => {
-                return compare(actual[key], partialDeep(expected?.original?.[key]));
-            });
-        }
-
-        const iscontainingDeep = Object.keys(expected).some(key => key.endsWith("mockit__iscontainingDeep"));
-        if (iscontainingDeep) {
-            if (typeof actual !== "object") {
-                return hasher.hash(actual) === hasher.hash(expected.original);
-            }
-
-            if (Array.isArray(expected.original)) {
-                return expected.original.every((expectedValue, index) => {
-                    return compare(actual[index], containingDeep(expectedValue));
-                });
-            }
-
-            if (expected.original instanceof Map) {
-                return Array.from(((expected?.original as Map<any, any>) ?? []).entries()).every(([key, expectedVAlue]) => {
-                    return compare(actual.get(key), containingDeep(expectedVAlue));
-                });
-            }
-
-            if (expected.original instanceof Set) {
-                return Array.from(((expected?.original as Set<any>) ?? []).values()).every(expectedValue => {
-                    return Array.from(actual.values()).some(actualValue => compare(actualValue, containingDeep(expectedValue)));
-                });
-            }
-
-            return Object.keys(expected.original).every(key => {
-                return compare(actual[key], containingDeep(expected.original[key]));
-            });
-        }
-
-        const containsSchema = objectContainsSchema(expected);
-
-        if (containsSchema) {
-            if (Array.isArray(expected)) {
-                return expected.every((item, index) => {
-                    return compare(actual[index], item);
-                });
-            }
-
-            if (expected instanceof Map) {
-                return Array.from(expected.entries()).every(([key, value]) => {
-                    return compare(actual.get(key), value);
-                });
-            }
-
-            if (expected instanceof Set) {
-                return Array.from(expected.values()).every(value => {
-                    return Array.from(actual.values()).some(actualValue => compare(actualValue, value));
-                });
-            }
-
-            return Object.keys(expected).every(key => {
-                return compare(actual[key], expected[key]);
-            });
-        }
-    }
-
-    if (typeof actual !== typeof expected) {
-        return false;
-    }
-
-
-    return hasher.hash(actual) === hasher.hash(expected);
-}
-
-
-function objectContainsSchema(obj: any) {
-    if (typeof obj !== "object") {
-      return false;
-    }
-  
-    if (Array.isArray(obj)) {
-      return obj.some((item) => objectContainsSchema(item));
-    }
-
-    if (obj instanceof Map) {
-        return Array.from(obj.values()).some((value) => objectContainsSchema(value));
-    }
-
-    if (obj instanceof Set) {
-        return Array.from(obj.values()).some((value) => objectContainsSchema(value));
-    }
-  
-    const keys = Object.keys(obj);
-    if (keys.some((key) => key.endsWith("mockit__isSchema"))) {
-      return true;
-    }
-  
-    return keys.some((key) => {
-      if (typeof obj[key] === "object") {
-        return objectContainsSchema(obj[key]);
-      }
-  
-      return false;
-    });
-  }
-  
-function objectContainsPartial(obj: any) {
-    if (typeof obj !== "object") {
-        return false;
-    }
-
-    if (Array.isArray(obj)) {
-        return obj.some((item) => objectContainsPartial(item));
-    }
-
-    if (obj instanceof Map) {
-        return Array.from(obj.values()).some((value) => objectContainsPartial(value));
-    }
-
-    if (obj instanceof Set) {
-        return Array.from(obj.values()).some((value) => objectContainsPartial(value));
-    }
-
-    const keys = Object.keys(obj);
-    if (keys.some((key) => key.endsWith("mockit__isPartial"))) {
-        return true;
-    }
-
-    return keys.some((key) => {
-        if (typeof obj[key] === "object") {
-            return objectContainsPartial(obj[key]);
-        }
-
-        return false;
-    });
-}
